Refactor

+++ src/main/java/com/asusoftware/myTransporter/user/services/UserService.java
@@ -3,6 +3,7 @@ package com.asusoftware.myTransporter.user.services;
 import com.asusoftware.myTransporter.address.model.Address;
 import com.asusoftware.myTransporter.address.model.dto.AddressDto;
 import com.asusoftware.myTransporter.address.services.AddressService;
+import com.asusoftware.myTransporter.exceptions.UserNotFoundException;
 import com.asusoftware.myTransporter.user.mappers.UserDtoEntity;
 import com.asusoftware.myTransporter.user.mappers.UserProfileDtoEntity;
 import com.asusoftware.myTransporter.user.model.User;
@@ -12,6 +13,8 @@ import com.asusoftware.myTransporter.user.model.dto.UserDto;
 import com.asusoftware.myTransporter.user.model.dto.UserProfileDto;
 import com.asusoftware.myTransporter.user.repository.UserRepository;
 import lombok.AllArgsConstructor;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Sort;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
@@ -38,20 +41,19 @@ public class UserService {
         Address address = addressService.findAddress(createUserDto.getAddressDto());
         user.setAddress(address);
         if(createUserDto.getToken() != null && createUserDto.getUserRole().equals(UserRole.CLIENT)) {
-            User transporter = userRepository.findUserByToken(createUserDto.getToken());
-            if(transporter != null) {
+            User transporter = userRepository.findUserByToken(createUserDto.getToken())
+                    .orElseThrow(() -> new UserNotFoundException(String.format("The transporter with this token: %s not found", createUserDto.getToken())));
                 user.setFollowed(transporter);
                 List<User> followers = transporter.getFollowers();
                 followers.add(user);
                 transporter.setFollowers(followers);
                 userRepository.save(user);
                 return ResponseEntity.ok().body(userDtoEntity.userToDto(user));
-            }
         } else {
             userRepository.save(user);
             return ResponseEntity.ok().build();
         }
-        return ResponseEntity.badRequest().build();
+       // return ResponseEntity.badRequest().build();
     }
 
 
@@ -60,7 +62,13 @@ public class UserService {
     }
 
     public User findById(UUID id) {
-        return userRepository.findById(id).orElse(null);
+        return userRepository.findById(id).orElseThrow(() -> new UserNotFoundException(String.format("The user with Id: %s, not found", id)));
+    }
+
+    public ResponseEntity<List<UserDto>> findAllTransporterFollowers(UUID transporterId, int pageNumber) {
+        PageRequest pageRequest = PageRequest.of(pageNumber, 10, Sort.by("firstName").ascending());
+        List<UserDto> followers = userRepository.findTransporterFollowersWithPagination(transporterId, pageRequest).stream().map(userDtoEntity::userToDto).collect(Collectors.toList());
+        return ResponseEntity.ok().body(followers);
     }
 
     public void saveUser(User user) {
@@ -75,11 +83,35 @@ public class UserService {
                 .map(userDtoEntity::userToDto).collect(Collectors.toList());
     }
 
-    public void delete(UUID id) {
-        userRepository.deleteById(id);
+    /**
+     * Only for transporter
+     */
+    public ResponseEntity<String> removeFollowerById(UUID followerId, UUID transporterId) {
+        // Find the transporter
+        User transporter = findById(transporterId);
+        if(transporter != null) {
+            List<User> clients = transporter.getFollowers();
+            // Remove the client from client list of the transporter
+            clients.removeIf(user -> user.getId().equals(followerId));
+            transporter.setFollowers(clients);
+            saveUser(transporter);
+            userRepository.deleteById(followerId);
+            return ResponseEntity.ok().body("Size: " + clients.size());
+        }
+        return ResponseEntity.notFound().build();
     }
 
-    public UserProfileDto getUserProfile(UUID id) {
-        return userProfileDtoEntity.userProfileToDto(userRepository.findById(id).orElse(null));
+    public ResponseEntity<String> delete(UUID id) {
+        User user = findById(id);
+        if(user.getUserRole().equals(UserRole.CLIENT)) {
+           return removeFollowerById(id, user.getFollowed().getId());
+        }
+        userRepository.deleteById(id);
+        return ResponseEntity.ok().build();
+    }
+
+    public ResponseEntity<UserProfileDto> getUserProfile(UUID id) {
+        UserProfileDto userProfileDto = userProfileDtoEntity.userProfileToDto(userRepository.findById(id).orElseThrow(() -> new UserNotFoundException("The actual user, not found")));
+        return userProfileDto != null ? ResponseEntity.ok().body(userProfileDto) : ResponseEntity.notFound().build();
     }
 }
